♻️ Extract modules to sdk
