+++ crates/client-sdk/Cargo.toml
@@ -18,6 +18,18 @@ borsh = "1.5.6"
 serde = "1.0"
 serde_json = { version = "1" }
 tracing = "0.1"
+anymap = { version = "0.12.1" }
+syn = { version = "2.0.101" }
+quote = { version = "1.0.39" }
+opentelemetry = { version = "0.28" }
+rand = { version = "0.8.5" }
+futures = { version = "0.3.31" }
+tracing-subscriber = { version = "0.3", default-features = false, features = [
+  "env-filter",
+  "fmt",
+  "json",
+] }
+paste = { version = "1.0.15" }
 
 risc0-zkvm = { version = "2.0.0", default-features = false, optional = true }
 sp1-sdk = { version = "4.1.3", default-features = false, optional = true }
