+++ crates/client-sdk/src/utils/logger.rs
@@ -129,7 +129,7 @@ pub enum TracingMode {
 
 /// Setup tracing - stdout subscriber
 /// stdout defaults to INFO to INFO even if RUST_LOG is set to e.g. debug
-pub fn setup_tracing(conf: &crate::utils::conf::Conf, node_name: String) -> Result<()> {
+pub fn setup_tracing(log_format: String, node_name: String) -> Result<()> {
     let mut filter = EnvFilter::builder()
         .with_default_directive(LevelFilter::INFO.into())
         .from_env()?;
@@ -158,7 +158,7 @@ pub fn setup_tracing(conf: &crate::utils::conf::Conf, node_name: String) -> Resu
     }
 
     // Can't use match inline because these are different return types
-    let mode = match conf.log_format.as_str() {
+    let mode = match log_format.as_str() {
         "json" => TracingMode::Json,
         "node" => TracingMode::NodeName,
         _ => TracingMode::Full,
