+++ crates/client-sdk/src/utils/static_type_map.rs
