+++ src/indexer/contract_state_indexer.rs
@@ -10,12 +10,11 @@ use tracing::debug;
 
 use crate::{
     bus::{BusClientSender, BusMessage},
-    log_debug, log_error,
     model::{Blob, BlobTransaction, Block, CommonRunContext, Transaction, TransactionData},
-    module_bus_client, module_handle_messages,
     node_state::module::NodeStateEvent,
     utils::{conf::Conf, modules::Module},
 };
+use client_sdk::{log_debug, log_error, module_bus_client, module_handle_messages};
 
 #[derive(Debug, Clone)]
 pub struct CSIBusEvent<E> {
