+++ src/modules/prover.rs
@@ -2,9 +2,7 @@ use std::{fmt::Debug, sync::Arc};
 
 use crate::{
     bus::{BusClientSender, BusMessage},
-    log_error,
     model::CommonRunContext,
-    module_bus_client, module_handle_messages,
     node_state::module::NodeStateEvent,
     utils::modules::Module,
 };
@@ -12,6 +10,7 @@ use anyhow::{anyhow, Context, Result};
 use borsh::{BorshDeserialize, BorshSerialize};
 use client_sdk::{
     helpers::{risc0::Risc0Prover, ClientSdkProver},
+    log_error, module_bus_client, module_handle_messages,
     rest_client::NodeApiHttpClient,
     transaction_builder::TxExecutorHandler,
 };
