+++ src/node_state/module.rs
@@ -4,14 +4,14 @@ use super::metrics::NodeStateMetrics;
 use super::{NodeState, NodeStateStore};
 use crate::bus::{command_response::Query, BusClientSender, BusMessage};
 use crate::data_availability::DataEvent;
-use crate::log_error;
 use crate::model::Contract;
 use crate::model::{Block, BlockHeight, CommonRunContext, ContractName};
-use crate::module_handle_messages;
 use crate::utils::conf::SharedConf;
 use crate::utils::modules::{module_bus_client, Module};
 use anyhow::{Context, Result};
 use borsh::{BorshDeserialize, BorshSerialize};
+use client_sdk::log_error;
+use client_sdk::module_handle_messages;
 use hyle_model::{TxHash, UnsettledBlobTransaction};
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
