+++ src/single_node_consensus.rs
@@ -6,12 +6,12 @@ use crate::consensus::{CommittedConsensusProposal, ConsensusEvent, QueryConsensu
 use crate::genesis::GenesisEvent;
 use crate::mempool::QueryNewCut;
 use crate::model::*;
-use crate::module_handle_messages;
 use crate::utils::conf::SharedConf;
 use crate::utils::modules::module_bus_client;
 use crate::{model::SharedRunContext, utils::modules::Module};
 use anyhow::Result;
 use borsh::{BorshDeserialize, BorshSerialize};
+use client_sdk::module_handle_messages;
 use hyle_crypto::SharedBlstCrypto;
 use hyle_net::clock::TimestampMsClock;
 use staking::state::Staking;
@@ -211,9 +211,9 @@ mod tests {
     use crate::bus::dont_use_this::get_receiver;
     use crate::bus::metrics::BusMetrics;
     use crate::bus::{bus_client, SharedMessageBus};
-    use crate::handle_messages;
     use crate::utils::conf::Conf;
     use anyhow::Result;
+    use client_sdk::handle_messages;
     use hyle_crypto::BlstCrypto;
     use std::sync::Arc;
     use tokio::sync::broadcast::Receiver;
