+++ src/tcp_server.rs
@@ -2,9 +2,7 @@ use std::sync::Arc;
 
 use crate::{
     bus::{BusClientSender, BusMessage},
-    log_error,
     model::CommonRunContext,
-    module_handle_messages,
     utils::{
         conf::SharedConf,
         modules::{module_bus_client, Module},
@@ -13,10 +11,9 @@ use crate::{
 
 use anyhow::Result;
 use client_sdk::tcp_client::{codec_tcp_server, TcpServerMessage};
+use client_sdk::{log_error, module_handle_messages};
 use hyle_net::tcp::TcpEvent;
 use tracing::info;
-
-impl BusMessage for TcpServerMessage {}
 
 module_bus_client! {
 #[derive(Debug)]
