+++ src/utils/mod.rs
@@ -2,8 +2,5 @@
 pub mod bus_ws_connector;
 pub mod conf;
 pub mod integration_test;
-pub mod logger;
 pub mod modules;
-pub mod profiling;
 pub mod serialize;
-pub mod static_type_map;
